Fixes to get things to build and run
Builds and sucessfully runs with emscripten 2.0.23 (running in the Docker
image that they provide: https://emscripten.org/docs/getting_started/downloads.html#using-the-docker-image)

Fix missing #include of <stdint.h> in memory.cpp that was causing the
native build to fail due to an undefined `uintptr_t` type.

Rename semaphore.h to posix_sem.h (which parallels the posix_sem.cpp file
that seems to use it). It was shadowing the semaphore.h that Emscripten
ships, which was leading to it complaining that `sem_timedwait` was not
defined when compiling its STL and atomics.

Fix linking failing due to to -lm and -lSDL being based in to wasm-ld. Done
by having an --enable-emscripten configure flag that in turns ends up
controlling a set of libraries that we want to exclude from the link step.
There may be an easier way to do this.

Add `-s FORCE_FILESYSTEM=1 -s TOTAL_MEMORY=536870912` as explicit linker
flags, otherwise the FS_createPreloadedFile calls would not work because it
was not exported, and there would not be enough memory. This is presumably
related to the previous issue, it looks like previous versions of Emscripten
would end up generating bitcode output, and _emafterbuild.sh would be used
to do the final "linking" and conversion to asm.js/wasm. However, that now
happens directly as part of `make` (BasilikII is a JavaScript file) and thus
these flags need to be set earlier.

Fix display buffer being malloc'ed with a 0 size (due to SDL_SetVideoMode
never being called on the EMSCRIPTEN_SAB code path, and thus the `s` surface
struct never being initialized, and haing its `stride` field be 0). This
manifested itself as illegal instructions after the cscGrayPage video
command, because it would write alternating 0xffffffff and 0x00000000 words
all over the heap.

Also default to debug flags/options being off.