Rebuild Games! disk image
The previous version had a SimCity 2000 folder in the trash, which
was taking up disk space. Just emptying the trash was not enough,
the data blocks were still in the disk image (which is not sparse)
and thus taking up room when gzipped.

Fixed by creating a new disk image (via the native version of Basilisk II
with the Basilisk II GUI helper, as described at
https://www.emaculation.com/doku.php/basiliskii_osx_setup) and copying
the files over, so that there would be no deleted data to worry about.

Compressed file size goes from ~23MB to ~16MB