idlewait support
Uses Atomics.wait to wait until there's input so that we don't spin as
aggessively when the machine is idle. This relies on Basilisk patching
(see 4f07113555ac03688155e6decaec0070f4332814) and makes it work on the
Emscripten code path.

Required tweaking the loop on the main thread to not acquire the input
lock if there was nothing to send, so that we would allow it to be held
longer.

There's still an 8ms timeout for the wait so that we give the screen a
chance to redraw for animations.

My initial attempt at this used Asyncify (https://emscripten.org/docs/porting/asyncify.html)
to allow a truly async emscripten_idle_wait function to be implemented, which
used setTimeout to yield to the browser event loop. However, that added
build complexity (and binary size increases), and ran into issues with
setTimeout throttling (to the browser, all JS execution is due to chained
setTimeout calls, and we run into the 4ms nested throttling described by
https://html.spec.whatwg.org/multipage/timers-and-user-prompts.html#timers:concept-task-2).

Atomics.wait does not yield to the event loop, but we're not relying on that
anyway, so this simpler approach is used.