Avoid extra memcopy of pixels when blitting
For 32-bit displays Screen_blit is Blit_Copy_Raw, which is just a
wrapper around memcopy from the_buffer to browser_pixels. However, since
we just take the destination buffer and copy it to the shared array
buffer, we can instead do that directly.

We can now budget 1ms less in IdleWait for blitting, which reduces idle
tab CPU use to ~30%