Separate input reading from video refreshing.
We were only reading the input before the video frame (60hz), which
was leading to delays in processing it. This was most visible when
processing of Ethernet packets (we use an input flag to trigger
the Ethernet interrupt). When using a BroadcastChannel ethernet
transport the mimium RTT between two tabs was a ~8ms.

Switch to reading the input in cpu_do_check_ticks, which invoked much
more frequently in the CPU emulation loop (~1000 Hz). The RTT goes down
to 1-2ms, and mouse input feels less laggy as well (this approximates
the approach that was used in PCE.js, see
https://jamesfriend.com.au/porting-pce-emulator-browser).