Fixes necessary to get SheepShaver to build natively for macOS/with Xcode after autconf was run for Emscripten
Makes it easier to have parellel native and Emscripten builds to compare
behavior.

The main issue was that the autoconf-generated config.h was taking
precedence over the hand-authored one in MacOSX/config/config.h. As
a blunt workaround we make the generated one #include the hand-authored
one when building for macOS.

Also includes a framework search path fix to get kpx_cpu to actually
build with Xcode (it was not able to find SDL2).