Initial Emscripten/WASM port of SheepShaver
Mostly mirrors the BasiliskII setup, as far as making autoconf changes to direct
things on the correct path, use a custom video implementation (instead of SDL or
X11), etc.

To share code with the BasiliskII implementation the same approach as the other
BasiliskII/SheepShaver ports is create, and symlinks from SheepShaver into the
BasiliskII directory are created. The video_js.cpp implementation needs a bunch
of #if SHEEPSHAVER branches to make the BasiliskII implementatio work for
SheepShaver too, it's modeled after video_sdl.cpp.

SheepShaver peculiarities that had to be taken into account:
- We can't rely on being able to intercept SIGSEGVs to handle low memory
  globals, we instead put Emscripten on the same code path as the macOS x86-64
  port (see 5065fb9d9504764a39ed471890d1ecdfc367bc31).
- Hardcode MAP_ANONYMOUS to being available when using mmap, otherwise we can't
  mmap the /dev/zero equivalent.
- We don't currently use WASM threads, so instead of a separate tick thread we
  periodically yield from the CPU emulator and trigger the VBL and tick
  interrupts

Does not include audio, input, ethernet, idlewait or other features, but is
enough to boot into the Finder given a Mac OS 9.0.4 image.

Also includes a bit of cleanup for both projects:
- Reduce the initial Emscriptem heap size from 512MB to 256MB -- we don't
  actually need that much.
- Remove unncessary manual config.h re-defines for BasiliskII (autoconf generats
  the correct values)

Updates mihaip/infinite-mac#34