Hook up input in SheepShaver
Mostly a matter of invoking the ReadJSInput function in CheckTicks(), though
I did make get invoked more frequently to match the latency improvement from
df44e317809a87acf71da54a6f9f439fdd675748.

The one tricky part was that mouse movement was not recognized (even
though clicks were, as was clipboard input). I enabled Mouse Keys and
that was similarly not working to move the mouse.

I instrumented the `CursorDeviceDispatch` trap call in adb.cpp (which ends
up being equivalent to `CursorDeviceMoveTo`), and it seemed to be correct.

Based on the parameters that it passes, `mouse_base + 4` is a `CursorDevicePtr`.
If we loook at the layout of that struct, it contains a `CursorDataPtr` at
offset 4, which in turn contains both `whereX`/`whereY` `Fixed` values (with the
desired position) as well as a `where` `Point` with the actual position.

By looking at those fields before and after the call, it looked like `whereX`
and `whereY` were being updated correctly, but `where` was not changing.
I then compared it with the behavior in the native build of SheepShaver, and
surprisingly, `where` was not changed at the end of the call either. However,
by the time of the next mouse move, it had been updated, so there must have
been another process that was updating it.

Reading through the original ADB Technote (https://developer.apple.com/library/archive/technotes/hw/hw_01.html)
it described CursorDeviceMoveTo as: "sets the absolute position of the cursor to
(absX, absY). The next time the cursor position is calculated, it will be
moved to this absolute location." It was unclear when this "next time" was,
but since I knew where in memory it was (via vm_do_get_real_address), I set
an LLDB watchpoint on it. The `CursorDeviceMoveTo` call cleared it to 0,
and then it was updated to a new value in a VBL task. I thus discovered that
I had not carried over the `VSLDoInterruptService` call when I made video_js.cpp
also build for SheepShaver.

Updates mihaip/infinite-mac#34