Switch Emscripten FPU emulation to mpfr
Matches the arm64 build, and avoids precision issues with the vanilla
double path in fpu_ieee.cpp.

There is a slowdown (Speedometer 4.0 FPU benchmark goes from 7.7 to 2.4),
but it's worth it to avoid quirks that prevent things like the scroll
thumbs in Mac OS 8 from positioning correctly.

Updates mihaip/infinite-mac#6