Add helper for monitoring the PRAM
Makes it easier to see when values are changed from user actions (and
thus if the default should be updated).

Uses this to change the default alert sound to Sosumi. Under
SheepShaver/Mac OS 9 this does not appear to be the source of truth,
there is also something in the file system (Sound Prefereces?) that
stores the selected sound (and overwrites the PRAM with it at startup).