Additional fixes for building on arm64
mpfr was building because emconfigure was setting HOST_CC to
/llvm-project/build/bin/clang, which can only target x86_64. Explicitly
set it to cc, which can build for arm64 (unclear why the configure
script was checking for that -- we're actually targeting wasm).

Also make sure that we clear old x86_64 artifacts before we build
the CPU emulator generator (make clean does not appear to remove them).