Fix building SheepShaver on ARM
We were passing in the actual CPU type to the configure script, which
was resulting in NATMEM_OFFSET being initialized, putting the emulator
on the vm_mac_acquire_fixed path, which is not actually supported
in Emscripten/WebAssembly.

Switch to passing in asmjs as the CPU type, which is the closest thing
to WebAssembly in the set of architectures known by the configure script.