Handle out of bounds memory reads in Basilisk II under Emscripten
The native build allows out of bounds reads (which presumably were OK
on actual hardware) via a SEGV handler. We don't have that capability
under WebAssembly/Emscripten (an "index out of bounds" exception is
thrown instead). We instead add a basic guard for out of bounds
reads in the memory reading functions. This will presumably slow them
down some (since there's a check/branch for every ready), but in practice
it does not appear to be noticeable (the Speedometer CPU score goes from
7.16 to 6.96).

Updates mihaip/infinite-mac#96