Fixes to build with Emscripten 3.13
Emscripten's mmap is stricter (as of emscripten-core/emscripten#17011)
and rejects non-0 starting addresses. Fix this by not updating the
next_address for Emscripten builds.

Tweak autoconf files (no longer need to force enable MAP_ANONYMOUS, but
SLIRP is being detected as supported, and we don't want it).

Don't assume that the emscripten build tools are in the environment
in _em_build_mpfr.sh.

Updates mihaip/infinite-mac#178